Extract to crate
