--- crates/init/templates/support/Cargo.toml
@@ -1,48 +1,0 @@
-[workspace]
-resolver = "2"
-members = ["app", "host", "methods"]
-exclude = ["methods/guest"]
-
-[workspace.package]
-version = "0.1.0"
-edition = "2021"
-
-[workspace.dependencies]
-app = { path = "./app" }
-host = { path = "./host" }
-program = { path = "../../program" }
-methods = { path = "./methods" }
-
-# alloy-primitives = { version = "0.8", default-features = false, features = [
-#     "rlp",
-#     "serde",
-#     "std",
-# ]}
-# alloy-sol-types = { version = "0.8" }
-# anyhow = { version = "1.0.86" }
-# actix-web = "4.10.2"
-# bincode = { version = "1.3.3" }
-# bytemuck = { version = "1" }
-# env_logger = "0.11.5"
-# hex = { version = "0.4.3" }
-# log = { version = "0.4" }
-# reqwest = { version = "0.12.15", features = ["json"] }
-# risc0-build = { version = "2.0.1", features = ["docker"] }
-# risc0-build-ethereum = { git = "https://github.com/risc0/risc0-ethereum", tag = "v2.0.0" }
-# risc0-ethereum-contracts = { git = "https://github.com/risc0/risc0-ethereum", tag = "v2.0.0" }
-# risc0-zkvm = { version = "2.0.0" }
-# risc0-zkp = { version = "2.0.0", default-features = false }
-# serde = { version = "1.0.208", features = ["derive", "std"] }
-# serde_json = "1.0.133"
-# fhe_rs = { package = "fhe", git = "https://github.com/gnosisguild/fhe.rs" }
-# fhe-traits = { git = "https://github.com/gnosisguild/fhe.rs" }
-# fhe-util = { git = "https://github.com/gnosisguild/fhe.rs" }
-# e3-compute-provider = { path = "../../../../crates/compute-provider" }
-# tokio = { version = "1.38", features = ["full"] }
-# rand = { version = "0.8.5" }
-# tracing-subscriber = { version = "0.3", features = ["env-filter"] }
-
-[profile.release]
-debug = 1
-lto = true
-
