--- crates/init/templates/support/host/Cargo.toml
@@ -1,24 +1,0 @@
-[package]
-name = "e3-support-host"
-version = { workspace = true }
-edition = { workspace = true }
-
-[dependencies]
-bincode = { workspace = true }
-bytemuck = { workspace = true }
-serde = { workspace = true }
-alloy-primitives = { workspace = true }
-alloy-sol-types = { workspace = true }
-anyhow = { workspace = true }
-log = { workspace = true }
-methods = { workspace = true }
-risc0-ethereum-contracts = { workspace = true }
-risc0-zkvm = { workspace = true }
-tokio = { workspace = true }
-e3-compute-provider = { git = "https://github.com/gnosisguild/enclave", branch = "hacknet" }
-fhe_rs = { workspace = true }
-fhe-traits = { workspace = true }
-fhe-util = { workspace = true }
-e3-user-program = { workspace = true }
-rand = { workspace = true }
-tracing-subscriber = { workspace = true }
