--- crates/init/templates/support/methods/guest/Cargo.toml
@@ -1,22 +1,0 @@
-[package]
-name = "guests"
-version = "0.1.0"
-edition = "2021"
-
-[workspace]
-
-[[bin]]
-name = "voting"
-path = "src/bin/voting.rs"
-
-[dependencies]
-alloy-primitives = { version = "0.8", default-features = false, features = ["rlp", "serde", "std"] }
-alloy-sol-types = "0.8"
-risc0-zkvm = { version = "2.0.0", default-features = false, features = ['std'] }
-e3-compute-provider = { git = "https://github.com/gnosisguild/enclave", rev = "hacknet" }
-e3-user-program = { path = "../../../../program" }
-bincode = "1.3.3"
-anyhow = "1.0.86"
-
-[profile.release]
-lto = "thin"
