+++ crates/support/.dockerignore
@@ -1,0 +1,2 @@
+/contracts
+/tests
