+++ crates/support/program/src/lib.rs
@@ -1,0 +1,17 @@
+use e3_bfv_helpers::decode_bfv_params_arc;
+use e3_compute_provider::FHEInputs;
+use fhe_rs::bfv::Ciphertext;
+use fhe_traits::{DeserializeParametrized, Serialize};
+
+/// CRISP Implementation of the CiphertextProcessor function
+pub fn fhe_processor(fhe_inputs: &FHEInputs) -> Vec<u8> {
+    let params = decode_bfv_params_arc(&fhe_inputs.params);
+
+    let mut sum = Ciphertext::zero(&params);
+    for ciphertext_bytes in &fhe_inputs.ciphertexts {
+        let ciphertext = Ciphertext::from_bytes(&ciphertext_bytes.0, &params).unwrap();
+        sum += &ciphertext;
+    }
+
+    sum.to_bytes()
+}
