+++ crates/support/scripts/build.sh
