+++ crates/support/scripts/run.sh
@@ -1,0 +1,10 @@
+# dev
+docker run -it \
+  -v $(pwd)/app:/app/app \
+  -v $(pwd)/host:/app/host \
+  -v $(pwd)/methods:/app/methods \
+  -v $(pwd)/program:/app/program \
+  -v $(pwd)/scripts:/app/scripts \
+  -v $(pwd)/Cargo.toml:/app/Cargo.toml \
+  -v $(pwd)/Cargo.lock:/app/Cargo.lock \
+  e3-support
