--- examples/basic/.enclave/support/app/Cargo.toml
@@ -1,14 +1,0 @@
-[package]
-name = "crisp-program"
-version = "0.1.0"
-edition.workspace = true
-
-[dependencies]
-actix-web.workspace = true
-env_logger.workspace = true
-serde.workspace = true
-serde_json.workspace = true
-tokio.workspace = true
-e3-compute-provider.workspace = true
-e3-support-host.workspace = true
-e3-program-client.workspace = true
