--- examples/basic/.enclave/support/host/README.md
@@ -1,3 +1,0 @@
-# Risc0 Host
-
-Exposes a function to run the computation found in `voting-core::fhe_processor` whilst also generating a risc0 proof assuming `risc0_dev_mode` is _not_ set.
