--- examples/basic/.enclave/support/methods/guest/README.md
@@ -1,12 +1,0 @@
-# Guest Programs
-
-Each file in the [`src/bin`](./src/bin) folder defines a program for the zkVM.
-We refer to the program running in the zkVM as the "[guest]".
-
-To learn more about writing guest programs, check out the zkVM [developer docs].
-For zkVM API documentation, see the [guest module] of the [`risc0-zkvm`] crate.
-
-[guest]: https://dev.risczero.com/terminology#guest
-[developer docs]: https://dev.risczero.com/zkvm
-[guest module]: https://docs.rs/risc0-zkvm/latest/risc0_zkvm/guest/index.html
-[`risc0-zkvm`]: https://docs.rs/risc0-zkvm/latest/risc0_zkvm/index.html
