--- examples/basic/.enclave/support/rust-toolchain.toml
@@ -1,4 +1,0 @@
-[toolchain]
-channel = "1.85"
-components = ["rustfmt", "rust-src"]
-profile = "minimal"
