--- examples/basic/.enclave/support/scripts/build.sh
@@ -1,1 +1,0 @@
-cd ./.enclave/support && cargo build
